Move DB calls into an API method and wrap them in a simple cache
