+++ package.json
@@ -12,9 +12,11 @@
     "@chakra-ui/react": "^1.5.2",
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.3.0",
+    "axios": "0.21.1",
     "basic-ftp": "^4.6.6",
     "framer-motion": "^4.1.5",
     "next": "10.1.3",
+    "node-cache": "5.1.2",
     "react": "17.0.1",
     "react-dom": "17.0.1"
   }
